+++ CHANGELOG.md
@@ -7,12 +7,19 @@ that can be found in the LICENSE file. -->
 > [!IMPORTANT]  
 > See the [Migration Guide](guides/migration_guide.md) for the details of breaking changes between versions.
 
-## 9.2.0
+## Unreleased
 
 ### New features
 
 - Introduce `AssetsChangeCallback` and `AssetsChangeRefreshPredicate`
   to help users act according to assets changes.
+
+### Fixes
+
+- Raise detailed negative range error.
+- Fix viewer confirm button predication.
+- Enlarge GIF gradients.
+- Fix potential paths assets count unexpected merging behaviors.
 
 ## 9.1.0